add mixin for check finish duel
